Add a new system for writing quickprint settings
This refactor helps out to prevent clutter in the gui.simpleMode file
by separating the actual configuration into a different file per machine
and also by using a list of dictionaries to determine which settings
should be used depending on the enabled options of the quickprint panel,
without the use of if/elif/else statements.
A setting can also be made into a callable function which can do more
processing than a static value.
This is the first step in fixing issue #80